1.增加了CocoaLumberjack日志框架 2.添加了进入前后台、打断事件的处理 3.添加了后台播放

+++ IdiotAVPlayer/IdiotAVPlayer/Resource.h
@@ -17,9 +17,9 @@ typedef NS_ENUM(NSInteger,ResourceType) {
 @interface Resource : NSObject
 
 @property (nonatomic, strong) NSURL *       requestURL;     //请求网址
-@property (nonatomic, assign) NSUInteger    requestOffset;  //请求起始位置
-@property (   atomic, assign) NSUInteger    cacheLength;    //缓冲长度
-@property (nonatomic, assign) NSUInteger    fileLength;     //文件长度
+@property (nonatomic, assign) long long     requestOffset;  //请求起始位置
+@property (   atomic, assign) long long     cacheLength;    //缓冲长度
+@property (nonatomic, assign) long long     fileLength;     //文件长度
 @property (nonatomic,   copy) NSString *    cachePath;      //缓存文件路径
 @property (nonatomic, assign) ResourceType  resourceType;   //资源类型
